refactor: Merge scroll hooks into one
