+++ hooks/index.tsx
@@ -2,8 +2,7 @@ export { useDebouncedCallback } from './use-debounced-callback'
 export { useHeadings } from './use-headings'
 export { useNewsletter } from './use-newsletter'
 export { useReducedMotion } from './use-reduced-motion'
-export { useScrollDirection } from './use-scroll-direction'
-export { useScrollProgress } from './use-scroll-progress'
+export { useScroll } from './use-scroll'
 export { useScrollToHeading } from './use-scroll-to-heading'
 export { useTheme } from './use-theme'
 export { useDocumentDimensions } from './use-document-dimensions'
