--- hooks/use-scroll-direction.tsx
@@ -1,38 +1,0 @@
-import { useEffect, useRef, useState } from 'react'
-import { getScrollDirection } from '@/lib'
-
-export function useScrollDirection() {
-  const [scrollDirection, setScrollDirection] = useState<'up' | 'down'>('up')
-  const animationFrame = useRef<number>(0)
-
-  useEffect(() => {
-    let lastScrollY = window.scrollY
-    let ticking = false
-
-    const updateScrollDirection = () => {
-      const currentScrollY = window.scrollY
-
-      setScrollDirection(getScrollDirection(currentScrollY, lastScrollY))
-
-      lastScrollY = currentScrollY
-      ticking = false
-    }
-
-    const onScroll = () => {
-      if (!ticking) {
-        animationFrame.current = window.requestAnimationFrame(
-          updateScrollDirection
-        )
-        ticking = true
-      }
-    }
-
-    window.addEventListener('scroll', onScroll)
-    return () => {
-      window.cancelAnimationFrame(animationFrame.current)
-      window.removeEventListener('scroll', onScroll)
-    }
-  }, [])
-
-  return scrollDirection
-}
