+++ types/index.tsx
@@ -18,6 +18,10 @@ export {
   type SearchPageProps,
   type Pages
 } from './page'
-export { type ScrollProgressSelector } from './scroll'
+export {
+  type ScrollDirection,
+  type ScrollProgressSelector,
+  type ScrollState
+} from './scroll'
 export { type Theme } from './theme'
 export { type TocHeadingNode } from './toc'
