+++ types/scroll.tsx
@@ -1,1 +1,6 @@
 export type ScrollProgressSelector = 'article' | 'html'
+export type ScrollDirection = 'up' | 'down'
+export type ScrollState = {
+  direction: ScrollDirection
+  progress: number
+}
